pyproject.toml
[project]
name = "python-project-svg-file"
version = "0.1.0"
description = ""
authors = [
    {name = "Kélian PONS",email = "kelian.pons@centrale.centralelille.fr"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "marimo (>=0.16.5,<0.17.0)",
    "numpy (>=2.3.3,<3.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "ruff (>=0.13.3,<0.14.0)"
]

[tool.poetry]
packages = [{include = "python_project_svg_file", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
